Minor fixes to samtools-enumerator
Ignore-this: d57c8824c462463efc73ce588625c80f

darcs-hash:20110401233038-fd04d-1b4b209e1631f8d8d4d9269c10298cf1cd1da0f0

+++ samtools-enumerator/src/Bio/SamTools/Enumerator.hs
@@ -14,6 +14,7 @@ import Control.Exception (bracket, bracket_, finally)
 import Control.Monad
 import Control.Monad.IO.Class
 import qualified Data.ByteString.Char8 as BS
+import Data.Int (Int64)
 import Data.List
 import Data.Maybe
 
@@ -54,13 +55,13 @@ enumQuery q = loop
         loop step = E.returnI step
         
 -- | Enumerate over the reads in a region from an indexed BAM file input handle
-enumIndexRegion :: BamIndex.IdxHandle -> Int -> (Int, Int) -> E.Enumerator Bam.Bam1 IO a
+enumIndexRegion :: BamIndex.IdxHandle -> Int -> (Int64, Int64) -> E.Enumerator Bam.Bam1 IO a
 enumIndexRegion h tid bnds step = E.Iteratee $ do 
   q <- liftIO $ BamIndex.query h tid bnds 
   E.runIteratee $ enumQuery q step
 
 -- | Enumerate over the reads in a region from a sorted, indexed BAM file
-enumBamRegion :: FilePath -> BS.ByteString -> (Int, Int) -> E.Enumerator Bam.Bam1 IO a
+enumBamRegion :: FilePath -> BS.ByteString -> (Int64, Int64) -> E.Enumerator Bam.Bam1 IO a
 enumBamRegion inname seqname bnds step = E.Iteratee $ liftIO $ bracket (BamIndex.open inname) (BamIndex.close) $ \h -> do
   tid <- lookupTid $ BamIndex.idxHeader h
   E.runIteratee $ enumIndexRegion h tid bnds step
